Changed all objects in gem2mt module from float to double and from short to long
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@2156 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ Modules/Submods/particlearray.cpp
@@ -25,10 +25,10 @@ static double Rand = -1;
 TParticleArray* TParticleArray::pa;
 
 
-TParticleArray::TParticleArray( short nPTypes, short nProps,
-           short *aNPmean,
-           short (*aParTD)[6],
-           short *anPmin, short *anPmax,
+TParticleArray::TParticleArray( long int nPTypes, long int nProps,
+           long int *aNPmean,
+           long int (*aParTD)[6],
+           long int *anPmin, long int *anPmax,
            TNodeArray* aNodes ):
   ParT0(0), ParT1(0), anParts(0),
   anPTypes(nPTypes), anProps(nProps), nodes(aNodes)
